continued thoughts on separation of things and registry - now with proxies
